Add camera functionality
User can now create a new Spot via the overflow menu. This launches
the NewSpotActivity which allows the user to input new Spot data and
take a photo with a camera app (if available on the user's device).
The newly created Spot is then saved to the database and will display
in the master list.