Add map permissions for current location
In MapFragment, onMapReady, app asks for current location
permissions. If granted, the current location button will
appear on the map. If clicked, the map zooms in to the
user's current location. If the current location dot is
pressed, the NewSpotActivity is launched with the
coordinates of the current location. The permissions code
is not solid; still need to fix a lot. Also, the database
is now preloaded with 3 real spots.