Launch SpotDetailActivity on info window click
When the user clicks on the Spot marker's info window, the Spot's
detail card is launched.

The communication between the MapFragment and MainActivity via
the onFragmentInteractionListener interface is sloppy because
in order to pass the spot ID to the Activity, the Fragment gets
it from the snippet on the marker. A better implementation would
be to get the markerPosition from the marker clicked and pass
that to the Activity through the interface method, but this would
require changing the code by which the SpotDetailActivity gets
the intent extra (which is currently the Spot ID).