Change NewSpotActivity to insert blank Spot immediately on launch
This was a major design change to NewSpotActivity. When launched,
we immedately create a blank Spot, insert it into the database,
and immediately retrieve it back from the database and cache it
locally. This fixes and simplifies many issues we had before due
to having to build up a new Spot (from user input) before the
actual Spot existed in the database. Since the Spot didn't have
an ID assigned to it yet, we had to save all user input into a
temporary dummy Spot and then copy it to a real spot when the
user clicked the Create Spot button. This was especially
problematic for photo file paths (which are based on the Spot ID).

With this commit, we immediately create, insert, and retrieve a
blank Spot. Then, all further operations are actually just
updates to that Spot. Since the ID is assigned to the Spot by
Room when it is created and first inserted into the db, it is
straightforward to create file paths for the Spot's photos
based on the ID.