Add photo editing to SpotPhotoViewPagerAdapter
When the user chooses to edit a photo, an appropriate photo
editing app is launched and the photo is edited. ACTION_EDIT
is not defined to return at result, however, oddly,
onActivityResult() in EditSpotActivity is still called with a null
intent and a 0 result value (success is -1). However, the
requestCode still comes through correctly, so it is used to
switch to the EDIT_PHOTO case and update the spot. The update is
necessary for the Observer to be called.