Make Spot title editable directly in SpotDetailActivity
We've used a separate EditSpotActivity to make changes to a spot,
but it would be cleaner and require substantially less code if
we just make the Spot info editable right in SpotDetail. When the
user clicks a TextView, it'll pop up an AlertDialog and allow the
user to edit the info.