Launch SpotDetailActivity to create new spot too
Previously, to create a new Spot, we launched NewSpotActivity.
Now, we want to use a single Activity (SpotDetailActivity) for
Spot detail display, Spot editing, and Spot creation. This was
challenging because of the "Observer problem": if the Spot is
fetched by name from the database, the Observer (on the Spot)
is unable to properly keep fetching the updated Spot if the name
changes. The solution here is to treat name changes, when this
Activity is launched in "Spot creation mode", differently than
other changes: only the UI is updated with the name change until
the user presses the back button to leave the Activity, that is
when we do the actual name update in the database (which is OK
because we no longer need to "observe" the Spot and update the UI.