Implement list view/grid view toggle in SpotMasterFragment
The toggle works by creating both adapters, and then setting the
appropriate LayoutManager and adapter. This may not be the most
performant implementation, but it works for now. Later, we might
consider using nested Fragments; a Fragment for the list view
and one for the grid view, then we just swap those in and out of
the container View on user clicks.